whisper_finetune/requirements.txt
transformers>=4.45.0
accelerate
torch
librosa
soundfile
jiwer
pandas
numpy
scikit-learn
pyyaml
